feat: split analytics into separate files
